Add backend logs

+++ server/modules/groups.py
@@ -1,4 +1,5 @@
 from modules import users
+from modules import logs
 
 import uuid
 import json
@@ -10,7 +11,9 @@ GROUPS_DIR_PATH = "./data/groups/"
 
 def _get_group_content(group_id: str) -> dict | None:
     group_file_path = GROUPS_DIR_PATH + group_id + ".json"
-    if not os.path.exists(group_file_path): return None
+    if not os.path.exists(group_file_path): 
+        logs.error("Groups", f"Failed to get group content as file was not found `{group_file_path}` [{group_id}]")
+        return None
 
     with open(group_file_path, "r") as file:
         return json.load(file)
@@ -19,6 +22,16 @@ def _save_group_content(group_id: str, content: dict) -> None:
     group_file_path = GROUPS_DIR_PATH + group_id + ".json"
     with open(group_file_path, "w") as file:
         json.dump(content, file)
+
+
+def _purge_blank_groups() -> None:
+    """ Remove all groups with no members and managers. """
+    for file in os.listdir(GROUPS_DIR_PATH):
+        group_id = file.split(".")[0]
+        content = _get_group_content(group_id)
+        if not content["members"] and not content["managers"]:
+            logs.warn("Groups", f"Purging blank group: [{group_id}]")
+            os.remove(GROUPS_DIR_PATH + file)
 
 
 def create_group(name: str, owner_uuid: str) -> str:
@@ -32,6 +45,7 @@ def create_group(name: str, owner_uuid: str) -> str:
             "managers": [owner_uuid],
         }, file)
 
+    logs.info("Groups", f"User <{owner_uuid}> created new group: `{name}` [{group_id}]")
     return group_id
 
 
@@ -41,6 +55,7 @@ def delete_group(group_id: str, deleter_uuid: str) -> str | bool:
         content = json.load(file)
     
     if content["owner_uuid"] != deleter_uuid:
+        logs.error("Groups", f"User <{deleter_uuid}> tried to delete group [{group_id}] but is not a owner. (owner: <{content['owner_uuid']}>)")
         return "Only a group owner can delete group."
     
     for user_uuid in content["members"]:
@@ -48,7 +63,12 @@ def delete_group(group_id: str, deleter_uuid: str) -> str | bool:
     for user_uuid in content["managers"]:
         users.remove_group_from_user_list(user_uuid, group_id)
 
+    if not os.path.exists(group_file_path):
+        logs.error("Groups", f"Failed to delete group [{group_id}] by the owner <{content['owner_uuid']}> (group file no longer exists?)")
+        return "Group is already removed."
+
     os.remove(group_file_path)
+    logs.warn("Groups", f"Deleted group [{group_id}] by <{deleter_uuid}>")
     return True
 
 
@@ -57,20 +77,25 @@ def add_member_to_group(group_id: str, user_uuid: str) -> None:
     if user_uuid not in content["members"]:
         content["members"].append(user_uuid)
         _save_group_content(group_id, content)
+        logs.info("Groups", f"Added member <{user_uuid}> to group [{group_id}]")
 
 
 def remove_member_from_group(group_id: str, user_uuid: str) -> bool:
     content = _get_group_content(group_id)
 
     if user_uuid not in content["members"] and user_uuid not in content["managers"]:
+        logs.error("Groups", f"Failed to remove <{user_uuid}> from group [{group_id}]. User not found as member nor manager")
         return False
     
     if user_uuid == content["owner_uuid"]:
+        logs.warn("Groups", f"Group owner <{user_uuid}> left the group [{group_id}]. Deleting group...")
         return delete_group(group_id, user_uuid)
         
     if user_uuid in content["members"]:
+        logs.warn("Groups", f"Removed member <{user_uuid}> from group [{group_id}]")
         content["members"].remove(user_uuid)
     if user_uuid in content["managers"]:
+        logs.warn("Groups", f"Removed manager <{user_uuid}> from group [{group_id}]")
         content["managers"].remove(user_uuid)
     
     _save_group_content(group_id, content)
@@ -91,6 +116,7 @@ def promote_group_member(group_id: str, promoted_uuid: str, promoter_uuid: str) 
     content["managers"].append(promoted_uuid)
     
     _save_group_content(group_id, content)
+    logs.info("Groups", f"Promoted group member: <{promoted_uuid}> to a manager by: <{promoter_uuid}> in group [{group_id}]")
     return True
 
 
@@ -108,6 +134,7 @@ def demote_group_member(group_id: str, demoted_uuid: str, demoter_uuid: str) -> 
     content["members"].append(demoted_uuid)
     
     _save_group_content(group_id, content)
+    logs.info("Groups", f"Demoted group manager: <{demoted_uuid}> to a member by: <{demoter_uuid}> in group [{group_id}]")
     return True
 
 
@@ -155,6 +182,7 @@ def rename_group(group_id: str, new_name: str) -> None:
     content = _get_group_content(group_id)
     content["name"] = new_name
     _save_group_content(group_id, content)
+    logs.info("Groups", f"Renamed group [{group_id}] to: `{new_name}`")
     
         
 # Invitations
@@ -165,11 +193,15 @@ def add_group_invitation(user_uuid: str, group_id: str) -> bool | str:
 
     users.DB.execute("INSERT INTO invitations (user_uuid, group_id) VALUES (?, ?)", (user_uuid, group_id))
     users.db_conn.commit()
+
+    logs.info("Groups", f"Invited user <{user_uuid}> to group [{group_id}]")
     return True
 
 def delete_group_invitation(user_uuid: str, group_id: str) -> None:
     users.DB.execute("DELETE FROM invitations WHERE user_uuid=? AND group_id=?", (user_uuid, group_id))
     users.db_conn.commit()
+    
+    logs.info("Groups", f"Removed user invitation <{user_uuid}> to group [{group_id}] (either accepted or rejected)")
     return True
 
 def get_user_group_invitations(user_uuid: str) -> list[list[str, str]]:
